+++ Assets/com.gamedeveducation.aibehaviours/Runtime/CommonCore/Navigation/INavigationInterface.cs
@@ -8,6 +8,9 @@ namespace CommonCore
         bool IsPathfindingOrMoving(GameObject InMover);
         bool HasReachedDestination(GameObject InMover);
         bool StopMoving(GameObject InMover);
+
+        bool IsLocationReachable(Vector3 InStartPosition, Vector3 InDestinationPosition, float InSearchRange = -1.0f);
+
         Vector3 FindNearestNavigableLocation(GameObject InMover, Vector3 InSearchPoint, float InSearchRange);
     }
 }